# packages/api-types/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Mini Hedge Fund Desk",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/prices/latest/{instrument_id}": {
      "get": {
        "tags": [
          "market_data"
        ],
        "summary": "Get Latest Price",
        "operationId": "get_latest_price_api_v1_prices_latest__instrument_id__get",
        "parameters": [
          {
            "name": "instrument_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Instrument Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PriceSnapshot"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/prices/history/{instrument_id}": {
      "get": {
        "tags": [
          "market_data"
        ],
        "summary": "Get Price History",
        "operationId": "get_price_history_api_v1_prices_history__instrument_id__get",
        "parameters": [
          {
            "name": "instrument_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Instrument Id"
            }
          },
          {
            "name": "start",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "title": "Start"
            }
          },
          {
            "name": "end",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "title": "End"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PriceSnapshot"
                  },
                  "title": "Response Get Price History Api V1 Prices History  Instrument Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/fx/rates/{base}/{quote}": {
      "get": {
        "tags": [
          "market_data"
        ],
        "summary": "Get Fx Rate",
        "operationId": "get_fx_rate_api_v1_fx_rates__base___quote__get",
        "parameters": [
          {
            "name": "base",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Base"
            }
          },
          {
            "name": "quote",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Quote"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FXRateSnapshot"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/eod/run": {
      "post": {
        "tags": [
          "eod"
        ],
        "summary": "Trigger Eod Run",
        "description": "Trigger an EOD run for the authenticated user's fund.",
        "operationId": "trigger_eod_run_api_v1_eod_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunEODRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EODRunResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/eod/status/{business_date}": {
      "get": {
        "tags": [
          "eod"
        ],
        "summary": "Get Eod Status",
        "description": "Get the status of the latest EOD run for a business date.",
        "operationId": "get_eod_status_api_v1_eod_status__business_date__get",
        "parameters": [
          {
            "name": "business_date",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Business Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/EODRunResult"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Get Eod Status Api V1 Eod Status  Business Date  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/eod/history": {
      "get": {
        "tags": [
          "eod"
        ],
        "summary": "Get Eod History",
        "description": "List past EOD runs for the fund.",
        "operationId": "get_eod_history_api_v1_eod_history_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EODRunSummary"
                  },
                  "title": "Response Get Eod History Api V1 Eod History Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/corporate-actions/process": {
      "post": {
        "tags": [
          "corporate-actions"
        ],
        "summary": "Process Corporate Actions",
        "operationId": "process_corporate_actions_api_v1_corporate_actions_process_post",
        "parameters": [
          {
            "name": "portfolio_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Portfolio Id"
            }
          },
          {
            "name": "start_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Start Date"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "End Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProcessedAction"
                  },
                  "title": "Response Process Corporate Actions Api V1 Corporate Actions Process Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ActionType": {
        "type": "string",
        "enum": [
          "dividend",
          "stock_split",
          "reverse_split",
          "spinoff"
        ],
        "title": "ActionType"
      },
...
      "EODRunResult": {
        "properties": {
          "run_id": {
            "type": "string",
            "format": "uuid",
            "title": "Run Id"
          },
          "business_date": {
            "type": "string",
            "format": "date",
            "title": "Business Date"
          },
          "fund_slug": {
            "type": "string",
            "title": "Fund Slug"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/EODStepResult"
            },
            "type": "array",
            "title": "Steps"
          },
          "is_successful": {
            "type": "boolean",
            "title": "Is Successful"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "business_date",
          "fund_slug",
          "started_at",
          "steps",
          "is_successful"
        ],
        "title": "EODRunResult"
      },
      "EODRunSummary": {
        "properties": {
          "run_id": {
            "type": "string",
            "format": "uuid",
            "title": "Run Id"
          },
          "business_date": {
            "type": "string",
            "format": "date",
            "title": "Business Date"
          },
          "fund_slug": {
            "type": "string",
            "title": "Fund Slug"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "is_successful": {
            "type": "boolean",
            "title": "Is Successful"
          },
          "steps_completed": {
            "type": "integer",
            "title": "Steps Completed"
          },
          "steps_total": {
            "type": "integer",
            "title": "Steps Total"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "business_date",
          "fund_slug",
          "started_at",
          "is_successful",
          "steps_completed",
          "steps_total"
        ],
        "title": "EODRunSummary"
      },
      "EODStepName": {
        "type": "string",
        "enum": [
          "market_close",
          "price_finalization",
          "position_recon",
          "nav_calculation",
          "pnl_snapshot",
          "eod_risk",
          "fee_accrual",
          "capital_allocation",
          "performance_attribution"
        ],
        "title": "EODStepName"
      },
      "EODStepResult": {
        "properties": {
          "step": {
            "$ref": "#/components/schemas/EODStepName"
          },
          "status": {
            "$ref": "#/components/schemas/EODStepStatus"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "step",
          "status",
          "started_at"
        ],
        "title": "EODStepResult"
      },
      "EODStepStatus": {
        "type": "string",
        "enum": [
          "pending",
          "running",
          "completed",
          "failed",
          "skipped"
        ],
        "title": "EODStepStatus"
      },
...
      "FXRateSnapshot": {
        "properties": {
          "base_currency": {
            "type": "string",
            "title": "Base Currency"
          },
          "quote_currency": {
            "type": "string",
            "title": "Quote Currency"
          },
          "rate": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Rate"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "source": {
            "type": "string",
            "title": "Source"
          }
        },
        "type": "object",
        "required": [
          "base_currency",
          "quote_currency",
          "rate",
          "timestamp",
          "source"
        ],
        "title": "FXRateSnapshot",
        "description": "Immutable snapshot of a single FX rate."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PositionAdjustment": {
        "properties": {
          "instrument_id": {
            "type": "string",
            "title": "Instrument Id"
          },
          "quantity_delta": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Quantity Delta"
          },
          "cost_basis_adjustment": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Cost Basis Adjustment"
          },
          "cash_amount": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Cash Amount"
          }
        },
        "type": "object",
        "required": [
          "instrument_id",
          "quantity_delta",
          "cost_basis_adjustment",
          "cash_amount"
        ],
        "title": "PositionAdjustment"
      },
...
      "PriceSnapshot": {
        "properties": {
          "instrument_id": {
            "type": "string",
            "title": "Instrument Id"
          },
          "bid": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Bid"
          },
          "ask": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Ask"
          },
          "mid": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Mid"
          },
          "volume": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Volume"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "source": {
            "type": "string",
            "title": "Source"
          }
        },
        "type": "object",
        "required": [
          "instrument_id",
          "bid",
          "ask",
          "mid",
          "timestamp",
          "source"
        ],
        "title": "PriceSnapshot"
      },
      "ProcessedAction": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "action_id": {
            "type": "string",
            "title": "Action Id"
          },
          "instrument_id": {
            "type": "string",
            "title": "Instrument Id"
          },
          "action_type": {
            "$ref": "#/components/schemas/ActionType"
          },
          "ex_date": {
            "type": "string",
            "format": "date",
            "title": "Ex Date"
          },
          "status": {
            "$ref": "#/components/schemas/ProcessingStatus"
          },
          "adjustments": {
            "items": {
              "$ref": "#/components/schemas/PositionAdjustment"
            },
            "type": "array",
            "title": "Adjustments",
            "default": []
          },
          "processed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Processed At"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          }
        },
        "type": "object",
        "required": [
          "action_id",
          "instrument_id",
          "action_type",
          "ex_date",
          "status"
        ],
        "title": "ProcessedAction"
      },
      "ProcessingStatus": {
        "type": "string",
        "enum": [
          "pending",
          "processed",
          "failed",
          "skipped"
        ],
        "title": "ProcessingStatus"
      },
...
      "RunEODRequest": {
        "properties": {
          "business_date": {
            "type": "string",
            "format": "date",
            "title": "Business Date"
          }
        },
        "type": "object",
        "required": [
          "business_date"
        ],
        "title": "RunEODRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}